refactor: refactored expressions builder to have a file per expression type
